ast: Fix object corruption during safety reordering
The safety check was corrupting object and set values that contained
comprehension as object keys or set elements because the comprehension
values themselves were mutated in place. This change fixes the issue
by copying object/set values like we do in other places.

This change also removes the setExprIndices function which was also
mutating values inside of a visitor.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>