wasm: Add minimum malloc size and split threshold
To help reduce fragmentation we will now have a minimum of 16 byte
sized allocation block. In testing with JSON parsed objects we
generate a significant amount of 12 and 16 byte allocations for the
container structs for opa values, the size chosen should fit the
majority of them. In addition we will not split free blocks unless
the remainder is big enough to make additional allocations.

Testing with a relatively large (6MB) nested JSON object being parsed
this cuts the length of our free block list from a few thousand down
to zero. The time it takes to parse it (on my machine) goes from about
30 seconds to a few hundred milliseconds.

Signed-off-by: Patrick East <east.patrick@gmail.com>