wasm: "Reset" the heap for malloc unit tests
A lot of the tests assume knowledge about the state of the free block
list and usage of the heap. This gets thrown off when >1 test is
messing around with the heap, and the ordering of tests isn't
guaranteed.

This changes to do a hacky reset to reset the heap to a starting
position at the current heap top. This simulates a "fresh" vm with
no pre-allocated memory but leaks anything up until then.

Signed-off-by: Patrick East <east.patrick@gmail.com>