wasm: Fix JSON parser to copy memory for strings and numbers
Previously the parser was constructing strings and numbers with
references to the input buffer memory. If that memory was freed after
the opa_json_parse() call, it would corrupt the strings and numbers
returned by the parse.

This commit just updates the parser to create a copy of the
string/number values. If this becomes a performance issue in the
future, we can introduce an optional API for callers that promise NOT
to free the input buffer.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>