wasm: malloc with multiple free lists.
Split the single free list to many: each list holds chunks of specific
size, and therefore, while searching for a free block, identifying the
right list is cheap. To reduce fragmentation all but the list with
largest chunks holds fixed size blocks.

The number of free lists and chunk sizes (currently 4, 8, 16, 64, and
128+ bytes) warrant proper tuning.

Signed-off-by: Teemu Koponen <koponen@styra.com>