wasm: Add native support for walk() built-in and relations
This commit adds native support for the walk() built-in inside the
C library. As part of this change, the planner has been updated to
treat relational built-ins differently from normal built-in
functions. Specifically, the planner assumes the relation is
implemented by a function that generates the full extent of the
relation. The planner then generates a scan across the relation
output, unifying each value with the output operand.

Fixes #2771

Signed-off-by: Torin Sandall <torinsandall@gmail.com>