topdown/strings: make strings.replace_n sort its pattern object keys
In the existing implementation, looping over the map was indeterministic,
and hence the result of subsequent evaluations of a string.replacen_n call
with overlapping patterns would be, too.

Now, the builtin will sort its keys before building the strings.Replacer.

Before, the ordering of overlapping patterns wasn't under the user's control,
and after, it still isn't. But it's deterministic.

Signed-off-by: Stephan Renatus <srenatus@chef.io>