ci: Prevent commit loop if build is non-deterministic
The "post-merge" flow would trigger whenever a commit is pushed, which
includes when we auto-generate the wasm binaries. If (for whatever
reason) that generate step was non-deterministic it would potentially
mean the CI looping and adding commits until we manually intervene.

This changes to prevent the step from being able to commit on top of
another auto-generated commit.

Signed-off-by: Patrick East <east.patrick@gmail.com>