wasm: Fix planning of ordered rules
Previously the planner generated an instruction at the end of each
block to check if the return value was defined. If for any reason the
block were to set the return value but then fail before reaching the
end, the check would be skipped and execution would continue to the
next block that in turn could attempt to set the return value causing
a runtime error.

This commit modifies the planner to instead generate the return value
check at the beginning of each block except the first. This ensures
that blocks short-circuit (as expected for ordered rules) but it also
ensures that failures in previous block do not accidentally cause a
fallthrough to later blocks.

This issue was observed with a policy that happened to generate a scan
over the virtual _and_ base document trees.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>