internal/wasm/sdk: Bump default minimum memory size
The older size was 128kb, but the binaries generated by OPA are now
something closer to 600kb so we would almost always need to expand
the size. We will now start with a bigger minimum of 1mb to have a
better chance of fitting policies into memory without having to grow
right away.

Signed-off-by: Patrick East <east.patrick@gmail.com>