build: New flag for wasmer dependencies
We would previously gate building the wasm resolver (and in turn pull
in the wasmer stuff) based on whether or not cgo was enabled. However
for library use-cases this isn't ideal as someone might be using cgo
for an unrelated reason, and not wish to enabled the wasmer code in
opa.

There is now a new go tag for the files called "opa_wasmer" and a
makefile flag WASMER_ENABLED (default on) which will trigger it. This
means any libraries building OPA that want to have it enabled will
need to specify this tag with like `-tags=opa_wasmer ` or similar when
building.

Signed-off-by: Patrick East <east.patrick@gmail.com>