.github: Update post tag workflow to use env files
The post tag workflow used the set-env command to
set the TAG_NAME env variable. The set-env command is
now disabled due to a security vulnerability in the GitHub Actions
runner that can allow environment variable and path injection in
workflows that log untrusted data to STDOUT. For more information see:
https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/

This change updates the workflow to remove
usage of the set-env command to use environment files
instead.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>