Update tutorial to work running as non-root (#2967)
By default, in an OpenShift cluster, pods will run as a randomly
generated non-root user. That works fine with opa, but this example
attempts to bind to port 443. Moving to bind at 8443 allows the pod to
run as a non-root user successfully (tested).

By doing this, we can also use the rootless image which is a more secure
deployment method for all Kubernetes clusters.

Signed-off-by: James Hewitt <james.hewitt@uk.ibm.com>