wasm: have planner optimize case without rules in leaves (do nothing)  (#3089)
In our previous optimization change, #3058, we've been falling back to the
old behaviour (with introduces cross-products) whenever our optimization
didn't seem to apply.

This PR extends that a little: when there are only leaves (matching our
optimization lookup) that do not have rules (i.e. empty packages), we decide
that nothing package-related is to be done at all. So, we'll go on looking at
base data only.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>