wasm: Do not emit unreachable instructions after opa_abort calls
The unreachable instructions cause a crash in the Go runtime on
macOS. It's unclear why they are ever executed after calling
opa_abort/opa_runtime_error but this was observed to fix the issue.

In the future it would be nice to get rid of the panics in the
Go-defined functions like opa_abort. Instead of panicking those
functions should return traps. The problem is that currently returning
a trap requires closing over a store related object which introduces a
memory leak.

Fixes #3168

Signed-off-by: Torin Sandall <torinsandall@gmail.com>